fix database

+++ app/Http/Controllers/Api/V1/ProjectController.php
@@ -8,6 +8,9 @@ use App\Http\Resources\ProjectCollection;
 use App\Http\Resources\ProjectResource;
 use App\Http\Resources\UserProjectCollection;
 use App\Models\Project;
+use App\Models\Team;
+use App\Models\TeamProject;
+use App\Models\User;
 use App\Models\UserProject;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -18,13 +21,20 @@ class ProjectController extends Controller
 {
     public function index(ProjectRequest $request)
     {
-        $project = QueryBuilder::for(Project::class)
-            ->allowedIncludes(['team', 'tasks','users','created_by_user'])
-            ->allowedFilters(['name'])
-            ->paginate($request->per_page ?? 10)
-            ->appends($request->all());
+        $user = Auth::user();
+        if(Auth::user()->hasRoleAdmin()){
+            $project = QueryBuilder::for(Project::class)
+                ->allowedIncludes(['teams', 'tasks','users','created_by_user'])
+                ->allowedFilters(['name'])
+                ->paginate($request->per_page ?? 10)
+                ->appends($request->all());
+            return $this->respondSuccess(new ProjectCollection($project));
+        }else{
+            $projects = $user->projects;
+            return $this->respondSuccess(new ProjectCollection($projects));
+        }
 
-        return $this->respondSuccess(new ProjectCollection($project));
+
     }
 
     /**
@@ -35,20 +45,42 @@ class ProjectController extends Controller
      */
     public function store(Request $request)
     {
+        $teams = $request->teams ?? null;
         $user = Auth::user();
         $data = $request->validate([
             'name' => ['required', 'string'],
-            'team_id' => [ 'required', 'integer', 'min:1', 'exists:teams,id'],
         ]);
         $data['created_by'] = $user->id;
+        if($request->user_id){
+            $data['created_by'] = $request->user_id;
+        }
 
         $project = Project::create($data);
 
-        if ($user) {
+        if($request->user_id){
             UserProject::create([
                 'project_id' => $project->id,
-                'user_id' => $user->id
+                'user_id' => $request->user_id
             ]);
+        }if ($teams) {
+            $usersInTeams = User::whereHas('teams', function ($query) use ($teams) {
+                $query->whereIn('team_id', $teams);
+            })->pluck('id');
+            $datas = [];
+            foreach ($usersInTeams as $item) {
+                $datas[] = [
+                    'user_id' => $item,
+                    'project_id' => $project->id
+                ];
+            }
+            $check = UserProject::insert($datas);
+
+            foreach ($teams as $team) {
+                TeamProject::create([
+                    'team_id' => $team,
+                    'project_id' => $project->id
+                ]);
+            }
         }
         return $this->respondCreated(
             new ProjectResource($project)
@@ -91,20 +123,26 @@ class ProjectController extends Controller
      */
     public function destroy(ProjectRequest $request, project $project)
     {
-        if ($project->delete()) {
-            return $this->respondOk(__('Project.delete_success', ['resource', 'Project ' . $project->name]));
+        $user = Auth::user();
+        if (Auth::user()->hasRoleAdmin() || $user->id = $project->created_by){
+            if ($project->delete()) {
+                return $this->respondOk(__('Project.delete_success', ['resource', 'Project ' . $project->name]));
+            }
         }
         return $this->respondError(__('Project.delete_fail', ['resource', 'Project ' . $project->name]));
     }
 
     public function getProject(Request $request)
     {
+        $user = Auth::user();
         $id = $request->team_id ?? null;
-        $query = Project::where('team_id', $id);
-        $project = QueryBuilder::for($query)
-            ->allowedFilters('name')
-            ->paginate($request->per_page ?? 10)
-            ->appends($request->all());
-        return $this->respondSuccess( new ProjectCollection($project));
+        if($id){
+            $team = Team::find($id);
+            $projects = $team->projects;
+            return $this->respondSuccess(new ProjectCollection($projects));
+        }else {
+            $projects = $user->projects;
+            return $this->respondSuccess(new ProjectCollection($projects));
+        }
     }
 }
